resolve: fix rename SourceCache -> Sources
